Add and rename card class
Due to the nature of a Magic card we must differentiate between a physical card we can hold in our hand and a card as an existance in the Multiverse. All instanceses of a named card have the same game play mechanics. But they differ in the "real" world where each card in a collection is its own physical object.