Implement function to add one or more cards to a collection
First we try to create a card template by using the name provided by the user. If a card with that name exists in the Multiverse we continue and ask for additional data. When all data in the template is valid it can be transferred to a card object by passing the template dictionary. Initially I thought of parsing the template object but realised it would be unnecessarily complicated when retrieving and creating a card from a collection.