Prevent card data to saved unnecessarily
Since the API handle "fuzzy" names but the local files are dependant on using "exact" names, there was an issue where card data would be saves twice in the same key. This was not a real problem, just very redundant. This is (at least temporarily) solved by once again checking the local data, but now for the "exact" card name. It does mean an extra API request, but it won't overwrite the same data and print a save message.